Mechanika szansy na rybe, karp, szczupak

+++ Assets/Scripts/FishingRod/FishingRodStats.cs
@@ -7,5 +7,5 @@ public class FishingRodStats : MonoBehaviour
     [Header("Power of rod - from 0 to 30000")]
     public int RodPower;
     [Header("Chance to hook something - from 1 to 100")]
-    public int chanceToHook;
+    public float chanceToHook;
 }
